+++ cmd/vega/commands/propose_upgrade.go
@@ -35,8 +35,8 @@ type ProposeUpgradeCmd struct {
 	config.OutputFlag
 	config.Passphrase `long:"passphrase-file"`
 
-	VegaReleaseTag     string `short:"v" long:"vega-release-tag" required:"true" description:"A valid vega core release tag for the upgrade proposal"`
-	UpgradeBlockHeight uint64 `short:"h" long:"height" required:"true" description:"The block height at which the upgrade should be made"`
+	VegaReleaseTag     string `description:"A valid vega core release tag for the upgrade proposal" long:"vega-release-tag" required:"true" short:"v"`
+	UpgradeBlockHeight uint64 `description:"The block height at which the upgrade should be made"   long:"height"           required:"true" short:"h"`
 }
 
 var proposeUpgradeCmd ProposeUpgradeCmd