+++ core/execution/common/equity_shares.go
@@ -39,8 +39,6 @@ type EquityShares struct {
 	totalPStake num.Decimal
 	// lps is a map of party id to lp (LiquidityProviders)
 	lps map[string]*lp
-	// the same map as above, but used at the end of the opening auction in successor markets to carry over ELS
-	parentLPs map[string]*lp
 
 	openingAuctionEnded bool
 }
@@ -52,7 +50,6 @@ func NewEquityShares(mvp num.Decimal) *EquityShares {
 		totalPStake: num.DecimalZero(),
 		totalVStake: num.DecimalZero(),
 		lps:         map[string]*lp{},
-		parentLPs:   map[string]*lp{},
 	}
 }
 
@@ -74,20 +71,11 @@ func (es *EquityShares) InheritELS(shares []*types.ELSShare) {
 			current.share = els.Share
 			// then treat the current commitment as a change to the commitment amount:
 			es.SetPartyStake(els.PartyID, update)
-		} else {
-			// this LP hasn't committed to this market (yet)
-			es.parentLPs[els.PartyID] = &lp{
-				stake:  els.SuppliedStake,
-				share:  els.Share,
-				vStake: els.VStake,
-				avg:    els.Avg,
-			}
 		}
 	}
 }
 
 func (es *EquityShares) RollbackParentELS() {
-	es.parentLPs = map[string]*lp{}
 	// get all current stakes
 	current := es.lps
 	// clear current state:
@@ -128,11 +116,6 @@ func (es *EquityShares) OpeningAuctionEnded() {
 	}
 	es.openingAuctionEnded = true
 	es.r = num.DecimalZero()
-	if len(es.parentLPs) > 0 {
-		// set the ELS in successor markets
-		// es.resolveParentLPs()
-		es.parentLPs = nil // ditch the old data
-	}
 }
 
 func (es *EquityShares) UpdateVStake() {
@@ -183,17 +166,6 @@ func (es *EquityShares) AvgTradeValue(avg num.Decimal) *EquityShares {
 
 // SetPartyStake sets LP values for a given party.
 func (es *EquityShares) SetPartyStake(id string, newStakeU *num.Uint) {
-	if !es.openingAuctionEnded && len(es.parentLPs) > 0 {
-		if parent, ok := es.parentLPs[id]; ok {
-			// we can only reach this point if we've inherited from a parent market, and this
-			// party, at that time, hadn't made a commitment yet
-			es.lps[id] = parent
-			// add the old ELS data to the running totals
-			es.totalPStake = es.totalPStake.Add(parent.stake)
-			es.totalVStake = es.totalVStake.Add(parent.vStake)
-			delete(es.parentLPs, id) // we've restored this now, so remove from this map
-		}
-	}
 	v, found := es.lps[id]
 	if newStakeU == nil || newStakeU.IsZero() {
 		if found {
@@ -378,54 +350,4 @@ func (es *EquityShares) SetCPShares(shares []*types.ELSShare) {
 		es.totalVStake = es.totalVStake.Add(lp.vStake)
 		es.lps[share.PartyID] = &lp
 	}
-}
-
-func (es *EquityShares) InheritELSBatch(shares []*types.ELSShare) {
-	// the simple way is to just keep track of all previous ELS data
-	// once we leave opening auction, we call `resolveParentLPs` and work it all out
-	// of course, we could instead check the map of parties who already provided liquidity
-	// and do what we need to do for those parties here, and add the remainder to the partentLP map
-	for _, els := range shares {
-		lp := lp{
-			stake:  els.SuppliedStake,
-			share:  els.Share,
-			vStake: els.VStake,
-			avg:    els.Avg,
-		}
-		es.parentLPs[els.PartyID] = &lp
-		// es.totalPStake = es.totalPStake.Add(els.SuppliedStake)
-		// es.totalVStake = es.totalVStake.Add(els.VStake)
-	}
-	if es.openingAuctionEnded {
-		es.resolveParentLPs()
-	}
-}
-
-func (es *EquityShares) resolveParentLPs() {
-	for id, lp := range es.parentLPs {
-		if sLP, ok := es.lps[id]; ok {
-			// ok, this party carries over their ELS, but the amounts may differ, first check if the amount is different:
-			if sLP.stake.Equal(lp.stake) {
-				// physical stake is identical, let's update the vStake total:
-				es.totalVStake = es.totalVStake.Sub(sLP.vStake).Add(lp.vStake)
-				// and update the vStake and avg
-				sLP.vStake = lp.vStake
-				sLP.avg = lp.avg
-			} else {
-				// now this one is a bit trickier -> we should swap out the new commitment, and treat it as a change in commitment:
-				// step 1: update the the totals:
-				es.totalVStake = es.totalVStake.Sub(sLP.vStake).Add(lp.vStake)
-				// remove the new physical stake, add the old
-				es.totalPStake = es.totalPStake.Sub(sLP.stake).Add(lp.stake)
-				newCommitment, _ := num.UintFromDecimal(sLP.stake)
-				es.lps[id] = lp // assign it the old LP
-				// now apply the different stake as though it's an amendment of the old one
-				es.SetPartyStake(id, newCommitment)
-			}
-		}
-	}
-	// recalculate all ELS
-	es.updateAllELS()
-	// clear the data
-	es.parentLPs = map[string]*lp{}
 }