+++ core/execution/config.go
@@ -33,12 +33,12 @@ const (
 type Config struct {
 	Level encoding.LogLevel `long:"log-level"`
 
-	Matching   matching.Config   `group:"Matching" namespace:"matching"`
-	Risk       risk.Config       `group:"Risk" namespace:"risk"`
-	Position   positions.Config  `group:"Position" namespace:"position"`
+	Matching   matching.Config   `group:"Matching"   namespace:"matching"`
+	Risk       risk.Config       `group:"Risk"       namespace:"risk"`
+	Position   positions.Config  `group:"Position"   namespace:"position"`
 	Settlement settlement.Config `group:"Settlement" namespace:"settlement"`
-	Fee        fee.Config        `group:"Fee" namespace:"fee"`
-	Liquidity  liquidity.Config  `group:"Liquidity" namespace:"liquidity"`
+	Fee        fee.Config        `group:"Fee"        namespace:"fee"`
+	Liquidity  liquidity.Config  `group:"Liquidity"  namespace:"liquidity"`
 }
 
 // NewDefaultConfig creates an instance of the package specific configuration, given a