+++ datanode/ratelimit/config.go
@@ -7,11 +7,11 @@ import (
 )
 
 type Config struct {
-	Enabled bool              `long:"enabled" description:"Enable rate limit of API requests per IP address. Based on a 'token bucket' algorithm"`
-	Rate    float64           `long:"rate" description:"Refill rate of token bucket; maximum average request rate"`
-	Burst   int               `long:"burst" description:"Size of token bucket; maximum number of requests in short time window"`
-	TTL     encoding.Duration `long:"ttl" description:"Time after which inactive token buckets are reset"`
-	BanFor  encoding.Duration `long:"banfor" description:"If IP continues to make requests after passing rate limit threshold, ban for this duration. Setting to 0 seconds disables banning."`
+	Enabled bool              `description:"Enable rate limit of API requests per IP address. Based on a 'token bucket' algorithm"                                              long:"enabled"`
+	Rate    float64           `description:"Refill rate of token bucket; maximum average request rate"                                                                          long:"rate"`
+	Burst   int               `description:"Size of token bucket; maximum number of requests in short time window"                                                              long:"burst"`
+	TTL     encoding.Duration `description:"Time after which inactive token buckets are reset"                                                                                  long:"ttl"`
+	BanFor  encoding.Duration `description:"If IP continues to make requests after passing rate limit threshold, ban for this duration. Setting to 0 seconds disables banning." long:"banfor"`
 }
 
 func NewDefaultConfig() Config {